+++ rawbuilder/mocker/__init__.py
@@ -1,1 +1,0 @@
-from .Mocker import Mocker